+++ Sources/KeyboardKit/_Deprecated/Keyboard+Deprecated.swift
@@ -9,10 +9,30 @@ public extension Keyboard {
     typealias KeyboardState = State
 }
 
+public extension KeyboardSettings {
+
+    @available(*, deprecated, renamed: "Keyboard.SettingsLink")
+    typealias Link = Keyboard.SettingsLink
+}
+
 public extension View {
     
     @available(*, deprecated, renamed: "keyboardState(_:)")
     func withEnvironment(fromState state: Keyboard.State) -> some View {
         self.keyboardState(state)
+    }
+
+    @available(*, deprecated, message: "Use the non-controller based modifier instead.")
+    func keyboardState(
+        from controller: KeyboardInputViewController
+    ) -> some View {
+        self.keyboardState(controller.state)
+    }
+
+    @available(*, deprecated, renamed: "keyboardState(from:)")
+    func withEnvironment(
+        fromController controller: KeyboardInputViewController
+    ) -> some View {
+        self.keyboardState(from: controller)
     }
 }